pyproject.toml
[tool.poetry]
name = "alembic-test"
version = "0.1.0"
description = ""
authors = ["chacachien <nguyentienphatbl123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
sqlmodel = "^0.0.16"
uvicorn = "^0.29.0"
fastapi = "^0.110.1"
pytz = "^2024.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
pytest = "^8.1.1"
trio = "^0.25.0"
anyio = "^4.3.0"
pytest-asyncio = "^0.23.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
logging = "^0.4.9.6"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
